블루프린트 사용 app.py 정리
