+++ templates/login.html
@@ -28,7 +28,7 @@
                     </div>
                 </form>
             </div>
-            <p class="mt-4 text-gray-600 text-right">Don't have an account? <a href="{{ url_for('signup') }}" class="text-blue-500">Sign up</a></p>
+            <p class="mt-4 text-gray-600 text-right">Don't have an account? <a href="{{ url_for('signup.signup') }}" class="text-blue-500">Sign up</a></p>
         </div>
 
     </body>
