+++ templates/main.html
@@ -31,10 +31,10 @@
 
             
             <div class="border rounded p-4 mt-4 text-center">
-                <a class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600" href="{{ url_for('weekly_mission', weeknumber=1) }}">1주차 미션</a>
-                <a class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600" href="{{ url_for('weekly_mission', weeknumber=2) }}">2주차 미션</a>
-                <a class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600" href="{{ url_for('weekly_mission', weeknumber=3) }}">3주차 미션</a>
-                <a class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600" href="{{ url_for('weekly_mission', weeknumber=4) }}">4주차 미션</a>
+                <a class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600" href="">1주차 미션</a>
+                <a class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600" href="">2주차 미션</a>
+                <a class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600" href="">3주차 미션</a>
+                <a class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600" href="">4주차 미션</a>
             </div>
         </div>
     </body>
