+++ templates/signup.html
@@ -33,7 +33,7 @@
                     </div>
                 </form>
             </div>
-            <p class="mt-4 text-gray-600 text-right">Already have an account? <a href="{{ url_for('login') }}" class="text-blue-500">Login</a></p>
+            <p class="mt-4 text-gray-600 text-right">Already have an account? <a href="{{ url_for('login.login') }}" class="text-blue-500">Login</a></p>
         </div>
 
     </body>
